AnalysisTree refactor
Summary:
**Tl;dr**: Instead of generating a collection of cards and then grouping/sorting them at display time, we could store them in a tree which captures desired grouping/storing.

Key changes are in ax/analysis/analysis.py; other changes are either downstream type signature changes or test changes.

As we’ve worked on improving the Ax UI user experience wrt. Analyses we keep coming back to this idea of an “AnalysisCard Group”, “master card”, etc. Conceptually we keep wanting to return to this idea of AnalysisCards existing in context together with each other, even though there is no way to capture this in the existing data model.

This diff implements the following changes:
* Introduce new classes
  * AnalysisCardBase: holds a name and timestamp
  * AnalysisCardGroup: extends AnalysisCardBase, holds children AnalysisCardBases
  * AnalysisCard: same as before, but now extends AnalysisCardBase
* Remove metadata used for sorting from AnalysisCard (level, category, attributes)
  * See doc for plan for what to do about trial_index
* Chance AnalysisCard.compute to return AnalysisCardBase instead of list[AnalysisCard]
* Move logic necessary for rendering list[AnalysisCard] to AnalysisCardGroup
  * Remove logic for sorting list[AnalysisCard] before rendering
* Remove AnalysisBlobAnnotation, which will now be handled on the encoder/decoder level (implemented in next diff)
  * Also removed ErrorAnalysisCard since it was just a handle for setting AnalysisBlobAnnotation
* **Removed all callsites to _save_analysis_cards_to_db_if_possible**

Note that this diff removes the ability to store AnalysisCards, which will be re-implemented in a subsequent diff to be landed at the same time as this diff. It also comments out many tests which relied on examining AnalysisCards to function; these will be reworked when we rethink dispatch.

Reviewed By: mgarrard

Differential Revision: D75250277