Sort experiment data by trial_index, arm_name, and metrics (#3885)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3885

Arm order in metric results were displayed per the data order. This led to a random ordering of arms in the UI {F1977891605}

To fix this, we want to sort the arm ordering in data.
Data will be sorted by trail_index, then arm_name. Arm name will be sorted as 'custom_name' < '0_1' < '0_2' < '0_11' < '0_100'

Reviewed By: lena-kashtelyan, saitcakmak

Differential Revision: D74409276