Always use model predictions, not empirical best, for best-point (#3894)
Summary:

In the short time that Ax benchmarking has supported an inference trace, it has been a perennial source of confusion that there are two different ways to get the "best point" used for the inference trace: the model-recommended best point and the empirical best point. Combined with the "oracle trace," there are three different traces that might be used to evaluate a noisy problem. "Empirical best" -- that is `use_model_predictions_for_best_point=False` -- rarely makes sense to do. It is essentially choosing a point on the basis of noise even if the model is much smarter than that.

This diff:
* Removes the option `use_model_predictions_for_best_point` from `BenchmarkMethod`

Differential Revision: D76156686