Remove SearchSpaceToFloat transform (#3897)
Summary:

Removes `SearchSpaceToFloat` transform and replaces its only use case with `SearchSpaceToChoice + ChoiceToNumericChoice + RemoveFixed`.

This transform does not serve a real modeling purpose, and only helps in getting the data into an easy to process format, which can be handled by other existing transforms. Removing it reduces the amount of transform code that we have to maintain.

Differential Revision: D76275126