Remove SearchSpaceToFloat transform (#3897)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3897

Removes `SearchSpaceToFloat` transform and replaces its only use case with `SearchSpaceToChoice + ChoiceToNumericChoice + RemoveFixed`.

This transform does not serve a real modeling purpose, and only helps in getting the data into an easy to process format, which can be handled by other existing transforms. Removing it reduces the amount of transform code that we have to maintain.

Reviewed By: esantorella

Differential Revision: D76275126

fbshipit-source-id: de256607609623c9560ed423decd6aa9bea326e0