Remove IVW transform from Y_trans (#3900)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3900

`IVW` transform merges repeated observations of a metric in an `ObservationData` object, which requires the experiment to have duplicate observations for the metric for any given `(trial_index, arm_name)`. Based on some offline discussion, this is not a case we expect to encounter in an Ax experiment. This diff removes it from `Y_trans`, which is a subset of the default set of transforms we use across Ax.

Reviewed By: lena-kashtelyan

Differential Revision: D76343356

fbshipit-source-id: e178c0bce8d09bec14de7b9a2711edc5ca8c178b