Fix transforms used in get_tensor_converter_adapter (#3908)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3908

In D76275126, this was updated to use `SearchSpaceToChoice` transform, but that doesn't work for all cases since `ChoiceParameter` does not support more than 1000 values.

A simpler solution is to use `MBM_X_trans`, which is part of the default set of transforms Ax uses. If they work as default for any given experiment, they should work well for this utility as well.

Reviewed By: esantorella

Differential Revision: D76431509