Update MetadataToFloat tests to use an actual experiment
Summary:
Using manually constructed objects (in this case `Observation`) in tests is problematic, since they make it quite difficult to adapt tests to future changes. These objects often include data that cannot be extracted from an actual Ax experiment. In this case, the issue is having multiple `Observation` corresponding to the same `trial_index, arm_name` with different metadata that is not map keys. The `metadata` on `ObservationFeatures` is extracted from `candidate_metadata_by_arm_signature`, which is fixed for a given `trial_index, arm_name` pair (assuming this is not `MapData`, in which case there would be different ones for different map values).

This diff updates `MetadataToFloat` tests to use an `Experiment` and the data that is extracted from that experiment.

Differential Revision: D76611204