Implement Cast.transform_experiment_data (#3935)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3935

Supports transforming `ExperimentData` with `Cast` transform. For flattening the search space, we use `df.apply` to loop over the rows and update them using existing `HSS.flatten_observation_features` method. This could be worth re-visiting later to see if it could be made more efficient -- or it may not be necessary if we decide to stop flattening altogether based on the results of the internship project.

For `MapKeyToFloat`, the actual transform is no-op, since we have the map keys in `observation_data` and we can simply extract them from there in `Adapter`.

Background: As part of the larger refactor, we will be using `ExperimentData` in place of `list[Observation]` within the `Adapter`.
- The transforms will be initialized using `ExperimentData`. The `observations` input to the constructors may be deprecated once the use cases are updated.
- The training data for `Adapter` will be represented with `ExperimentData` and will be transformed using `transform_experiment_data`.
- For misc input / output to various `Adapter` and other methods, the `Observation / ObservationFeatures / ObservationData` objects will remain. To support these, we will retain the existing transform methods that service these objects.
- Since `ExperimentData` is not planned to be used as an output of user facing methods, we do not need to untransform it. We are not planning to implement`untransform_experiment_data`.

Reviewed By: esantorella

Differential Revision: D76627279

fbshipit-source-id: dcf056745f8a09b148f474f07a88b75206471f20