Allow loading of learning curves with NaNs and Infs in `TensorboardMetric._process_records_to_df`
Summary: This commit raises a warning if NaNs of Infs are present in the tensorboard data, instead of raising an exception. This makes sure that the later transform and modeling stages have access to the most amount of data, and allows them to deal with the non-finite part of the data separately, instead of removing all data of the associated metric from consideration.

Differential Revision: D77044321