ScaledParetoEarlyStoppingStrategy (#2211)
Summary:

This commit introduces `ScaledParetoEarlyStoppingStrategy` an early stopping strategy for multi-objective optimization problems. The strategy consists of scaling the observed objective toward the reference point, implicitly creating a scaled Pareto frontier. If a trial qualifies for early stopping, it is stopped if it does not dominate the scaled Pareto frontier.

Two different variations of the Pareto scaling, based on a constant uniform scaling and based on the percentiles of each objective, are highlighted in the following plot (N4871490):

{F1455464697}

The following is a schematic comparison to using logical operations to combine single-objective strategies:

{F1381511964}

See [**this document**](https://docs.google.com/document/d/1gdoJ9NfnDXckMDkZAkXqrDtz7U9D3PO04JaBv-mDp-o/edit?usp=sharing) for details.

Differential Revision: D53369314