Raise exception when `BatchTrial` is passed to `BaseEarlyStoppingStrategy` (#3967)
Summary:

This commit adds a check for `BatchTrials` to `BaseEarlyStoppingStrategy`, since they are currently not supported.

In particular, align_partial_results requires a 1-1 mapping between trial indices and arm names, which is not the case for batch trials.

If we didn't catch that there were multiple arms per trial, the interpolation code below would interpolate between data points from potentially different arms, as only the trial index is used to differentiate distinct data for interpolation.

Reviewed By: esantorella

Differential Revision: D77308661