Allow for best-point predictions in MF/MT settings
Summary: Context: Ax recommends a best point in multi-task and multi-fidelity scenarios, even though we should not expect the recommendations to be very good. Let's surface those recommendations in benchmarking rather than putting in NaNs so that we can measure any improvements we make to best- point functionality in such settings.

Differential Revision: D76634560