Allow for best-point predictions in MF/MT settings (#3970)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3970

Context: Ax recommends a best point in multi-task and multi-fidelity scenarios, even though we should not expect the recommendations to be very good. Let's surface those recommendations in benchmarking rather than putting in NaNs so that we can measure any improvements we make to best- point functionality in such settings.

Reviewed By: Balandat

Differential Revision: D76634560

fbshipit-source-id: 12ea3b67474f9d53bf8f8ee68d8238358117a044