Allow loading of learning curves with NaNs and Infs in `TensorboardMetric._process_records_to_df` (#3949)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3949

This commit raises a warning if NaNs of Infs are present in the tensorboard data, instead of raising an exception. This makes sure that the later transform and modeling stages have access to the most amount of data, and allows them to deal with the non-finite part of the data separately, instead of removing all data of the associated metric from consideration.

Reviewed By: esantorella

Differential Revision: D77044321

fbshipit-source-id: aebcbd89401691fa9c8960ce3fee3d57cfbc2903