Fix broken test, add validation (#3983)
Summary:

Context: A test was not running all cases due to a `return` in the middle. It was failing with a mysterious error about tensor dimensions. This was because the test uses `DeterministicGenerationStrategy` which only produces one-parameter parameterizations regardless of what problem it is used with. Separately, it turned out that passing too many parameters rather than too few would result in some parameters being silently dropped with no error.

This PR:
* Changes the test to use a Sobol GenerationStrategy
* Adds validation
* Stops silently dropping extra parameters

Reviewed By: Balandat

Differential Revision: D77673552