Improve sanitization of metric names in API (#3990)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3990

Become more lenient with where sanitization can happen, specifically by replacing occurances of "." and "/" when they appear after a valid Python variable name and before any alphanumeric character.

Note that we have to be careful not to oversanitize as some dots and slashes should mean something to the parser! Ideally we can get into a world in which Ax does not accept parameter or metric names which do not obey the same rules as Python variable names, which tend to cause trouble in unpredictable parts of our stack.

Differential Revision: D77699011