Update MergeRepeatedMeasurements to use arm name rather than hash (#3980)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3980

The transform was using the hash of parameterization as the key for `arm_to_merged`. This diff replaces it with `arm_name` instead. The motivation is to avoid the need to compute the hash when adapting the transform to support `ExperimentData`.

Why I think it is ok to rely on `Observation.arm_name`, which is technically an optional field: `transform_observations` is used to transform the training data or cv training data in `Adapter`. When we extract the `Observation` from experiment, these all have `arm_name`. As written, the transform drops any observations that weren't in the data used to construct it anyways, so anything that doesn't fit the pattern would be dropped in either case.

Reviewed By: SebastianAment

Differential Revision: D77664239

fbshipit-source-id: db3e735750b93af369bab94982c21d317cb2cddd