Remove `TFidelityTrialEvaluation`; typing improvements (#4010)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4010

Context: Ax supported attaching trial evaluations in several formats which can then be parsed as regular data, map data, or fidelity evaluations depending on the format. As of D77904185, I don't think fidelity valuations can be used, although it is possible that someone might find a way to do it. We could consider deprecating this to be safe, but I'm pretty sure this is unused.

This PR:
* Removes `TFidelityTrialEvaluation`, making it no longer one of the possibilities in `TEvaluationOutcome`
* Fixes some type checks

Reviewed By: saitcakmak

Differential Revision: D77935826

fbshipit-source-id: c759d535af284d5d67753f6990f367a7eadaa834