Remove log & logit-scale options from MapKeyToFloat / MetadataToFloat
Summary: The log / logit scale options in these transfroms necessitate applying the corresponding transform to the parameters & parameter values added by these transforms. This creates some complications when using the transforms with `ExperimentData`, which was discussed in detail in D78094423. Based on that discussion, we're opting to disallow log / logit scale in these generated parameters, which simplifies the setup.

Reviewed By: Balandat

Differential Revision: D78168504