Filter Trials for Sampled Points on Plots
Summary:
`SlicePlot` and `ContourPlot` currently do not filter for `TrialStatus` when plotting sampled points (x's on the plot). This means we plot points as "sampled", even when the Trial has failed and the arms were never run.

This diff adds filtering for `TrialStatus` to be in `STATUSES_EXPECTING_DATA` (completed, running, early stopped) so the sampled points are only from arms that have been run.

Differential Revision: D78353511