Filter Trials for Sampled Points on Plots (#4038)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4038

`SlicePlot` and `ContourPlot` currently do not filter for `TrialStatus` when plotting sampled points (x's on the plot). This means we plot points as "sampled", even when the Trial has failed and the arms were never run.

This diff adds filtering for `TrialStatus` to be in `STATUSES_EXPECTING_DATA` (completed, running, early stopped) so the sampled points are only from arms that have been run.

Reviewed By: Balandat, ItsMrLin

Differential Revision: D78353511

fbshipit-source-id: d7eba4c35e6479f1a0d276bdec0d1e63a3d9d90a