Add "quality" method to configure_generation_strategy
Summary:
Add a new option on the "method" parameter called "quality". This parameter controls the relative computational expense of computing the next candidate during optimization -- users can specify "quality" to indicate they would like Ax to generate the highest quality candidates it is able to, at the expense of slow runtime.

Under the hood this dispatches to SaasFullyBayesianSingleTaskGP with input warping

Differential Revision: D78428692