Add additional logging to Client
Summary:
In line with the logging present in AxClient.

**Looking for input on which of these seem worth keeping.**

Its not clear to me that each of these additional logging improve the user experience, specifically those that either:
* Simply note that a method has been called with certain parameters
* Log the return value

Logs that let the user know about some side effect not captured by the return value and not obvious by the fact that the method was called seem worth keeping, (ex. when a user calls "complete_trial" and not all metrics were present we log that we're actually going to mark the trial FAILED).

Differential Revision: D78183140