Add "quality" method to configure_generation_strategy (#4042)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4042

Add a new option on the "method" parameter called "quality". This parameter controls the relative computational expense of computing the next candidate during optimization -- users can specify "quality" to indicate they would like Ax to generate the highest quality candidates it is able to, at the expense of slow runtime.

Under the hood this dispatches to SaasFullyBayesianSingleTaskGP with input warping

Reviewed By: saitcakmak

Differential Revision: D78428692

fbshipit-source-id: 170ae8c7171497882e05c5b6914299a92828bd8e