Fix for the map key error when using BatchClient for experiments that are configured using PTSClinet
Summary:
With the way client is configured now, whenever configure_optmization is called it default overrides the metric type to mapmetric which would in turn change the default_data_type attribute type to MapData for the experiment. With this the encoder encodes this experiment thinking it's a mapdata but when the decoder decodes it results in map key error since originally that metric is not a map metric.

Consequently, we have decided to loosen up the grip for the decoder by using .get() method in python dictionaries and have a fall back default value whenever it tries to get the key thinking it's a map metric. We believe that this won't have an extreme adverse effect.

Reviewed By: mpolson64

Differential Revision: D78506131