T231841440 : Remove trace hover in contour plot
Summary:
Remove trace hover for all points in contour plot; add hover only for sampled points in Ax analysis.

The contour plot currently shows hover information for all points, which is not very informative. This change removes the trace hover for the contour plot and adds hover information exclusively for the sampled points.

Reviewed By: mgarrard

Differential Revision: D79206300