Update ExperimentData.convert_to_list_of_observations to support map data
Summary: Previously, conversion to list of observations was only allowed if the `observation_data` did not include any map keys in its index. Some CV test failures in D70220094 revealed the need for including the progression / map key in the `ObservationFeatures` metadata for it to be transformed as part of the CV test points. The fix for adding map key to metadata seemed pretty close to supporting the map key more generally in this method (i.e., when there are multiple rows of data for each arm), so this diff implements that.

Differential Revision: D79356456

Reviewed By: ltiao