Fix TransformToNewSQ.transform_experiment_data (#4081)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4081

The transform is only supposed to drop the status quo observations for trials other than the target trial. The data for the target trial is meant to remain unchanged. This can be inferred from line 204 that skips the transform for target trial and line 225-228 that only drops SQ observations if it is not the target trial. This behavior is also verified by `TransferLearningAdapterTest.test_transform_to_new_sq` which is how I discovered the issue.

Reviewed By: SebastianAment

Differential Revision: D79290336

fbshipit-source-id: 9ab949c964cee151da80cf84b813d5f1eb971138