retain all metrics when changing the optimization config (#4092)
Summary:

It is a pretty confusing that metrics in the previous optimization config, but not the new optimization config are dropped when you change the optimization config. E.g. if you remove a constraint, you still likely want to track that metric. This change makes it so that metrics are not silently dropped when the optimization config changes.

Reviewed By: bletham

Differential Revision: D79475986