Updating the summarize method for user convenience
Summary:
#### Changes Summary:
This diff introduces an enhancement to the `summary` method in the `Analysis` class, allowing users to specify a subset of trial indices to include in the summary. This is achieved by adding a `trial_indices` parameter to the `to_df` method and modifying the underlying `lookup_data` call to filter for the specified trials.

**Key Changes:**

* Updated `to_df` method in `Experiment` class to accept `trial_indices` parameter
* Modified `lookup_data` call to filter for specified trials
* Added type hints for `trial_indices` parameter
* Improved code readability with added comments and whitespace

Reviewed By: mpolson64

Differential Revision: D79180301