Fix AnalysisCardGroup rendering in website (#4104)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4104

What can you say but "lol" 🙃

When building the website we execute tutorial notebooks (using papermill) then convert the ipynb to markdown using a mdx conversion script. The script breaks if we use our typical AnalysisCard._ipython_display_ so we have some special logic which simplifies rendering if we detect we're in papermill which flattens the group and renders each card individually.

We were exiting the *for loop* early instead of exiting the *if statement* because the `return` was indented one level too deep. lol.

Reviewed By: CristianLara

Differential Revision: D79726971

fbshipit-source-id: e0275ab9f427bb8c7dbd29978c96f6a713861248