Silence excessive warnings from ESS
Summary:
Looking at any tutorial with early stopping (ex: https://ax.dev/docs/next/tutorials/sklearn/) shows a ton of these warnings: `[WARNING 08-06 05:15:48] ax.early_stopping.strategies.base: No metric names specified. Defaulting to the objective metric(s).`

The docstring specifies this to be the behavior when metrics are not explicitly specifies. This warning does not need to be this intrusive, we can reduce it to debug level.

Differential Revision: D79740162