Silence excessive warnings from ESS (#4108)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4108

Looking at any tutorial with early stopping (ex: https://ax.dev/docs/next/tutorials/sklearn/) shows a ton of these warnings: `[WARNING 08-06 05:15:48] ax.early_stopping.strategies.base: No metric names specified. Defaulting to the objective metric(s).`

The docstring specifies this to be the behavior when metrics are not explicitly specifies. This warning does not need to be this intrusive, we can reduce it to debug level.

Reviewed By: ltiao

Differential Revision: D79740162

fbshipit-source-id: 1bd14c47912aa9697455f26ca5c2154a5b0afe58