Update MapKeyToFloat.transform_experiment_data to fill NaN progression (#4115)
Summary:

If any progression value is missing in `ObservationFeatures.metadata`, `MapKeyToFloat._transform_observation_feature` fills it with the upper bound of the corresponding parameter. This encodes the assumption that any data without a progression value corresponds to a completed trial (as a trial can be completed with `client.complete_trial` without providing progression).

The initial implementation of `transform_experiment_data` did not account for the possibility of NaN values in the corresponding index. This diff updates the method to check for NaN progression values and fill them with the target progression.

Reviewed By: ltiao

Differential Revision: D79750521