Update MetadataToParameter.transform_experiment_data to retain the original
Summary:
This transform was popping the corresponding element from the metadata before adding it as a feature. This is fine for the most part, but it doesn't really have a significant benefit over just leaving the original in the metadata.

The problem with popping the element from the metadata is when doing this with `ExperimentData`. Even if we `deepcopy` a pandas dataframe, it does not deepcopy the underlying dictionary values. If we then transform the copy, we end up removing the metadata from the original as well, which leads to errors when doing cross validation with an adapter that uses these transforms (since we end up transforming multiple copies, and only the first one has the metadata).

This diff updates the transfrom to not remove the element from metadata.

Differential Revision: D79805146