Add multi-model acquisition to MultiAcquisition (#4110)
Summary:

Augments MultiAcquisition to additionally handle the case where we want to generate from multiple models with the same botorch acquisition class. The collection of models from which to generate is created by Surrogate.

This requires making the MBM Acquisition aware of how many points are intended to be generated with this particular Acquisition. Note that we do construct a new Acquisition in every call to gen.

The diff does a refactor of `Acquisition.__init__` to move some of the logic into other methods that are called during `__init__`, and then can be re-called when we want to use the same Acquisition object to instantiate multiple botorch acquisition functions with different models each.

The `gen_metadata` entry that stores which model was used for the generation is now potentially a list of models, so I renamed the entry, both to be something a little more clear and also because there is a breaking change in the data type if one were trying to do some type of meta analysis anyway.

Differential Revision: D78560935