Updating docstrings for better code quality (#4119)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4119

This diff updates the docstrings in `client.py` and `summary.py` to improve code quality, making it easier for users to understand the purpose and usage of the functions. Specifically, the changes include:

*   Adding function argument explanations for `trial_indices` and `trial_status` in the `client.py` file.
*   Adding function argument explanations for `trial_indices`, `trial_status`, and `omit_empty_columns` in the `summary.py` file.

These updates enhance the overall clarity and readability of the codebase, ensuring that users can effectively utilize the functions and their arguments.

Reviewed By: mpolson64

Differential Revision: D79738530

fbshipit-source-id: 54b3f379399105eec0b4db4f070625280ea71fd7