Add reliable identifier to ModelConfig (#4101)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4101

ModelConfig has a name field that is optional. MBM Surrogate needs an identifier for model; currently uses name if present, or just drops the model if not when storing eval results. MBM Generator needs an identifier for model and uses name if present, or str(model_config) if not.

This cleans things up a bit by providing an identifier for ModelConfig that will always exist and so can be used in the place of name and clean up logic in the other parts of MBM. The identifier follows what was done in MBM Generator for tracking gen metadata: it is name if name exists, otherwise a string dump of the model config.

Reviewed By: sdaulton

Differential Revision: D78560884

fbshipit-source-id: 89a94dbeaf97d799e46a7d12f248cdf4382d3ea2