Add model sampling to Surrogate for batch generation (#4109)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4109

Adds a "models_for_gen(n)" method to Surrogate that will constructs a list of n models from which to generate n candidates. The list of models is constructed by sampling from the given ModelConfigs according to their model selection eval criterion (taken as a measure of model quality).

Reviewed By: sdaulton

Differential Revision: D77252555

fbshipit-source-id: 89c82ce62b47ee13569fe42cdff5427b4bdddf30