Move out of sample filter in ContourPlot
Summary:
See https://github.com/facebook/Ax/issues/4103 for motivation.

We need to check for search space membership when we sample from the surrogate to draw the contour plot, specifically because in experiments where the status quo arm is all None we were getting errors.

However, we dont actually care if points are out of sample is technically out of the search space due to parameter constraint violation (ex. if the value for a non-plotted parameter selected by select_fixed_value causes some parameter constraint to be violated we shouldnt mind and Ax should go ahead and plot regardless). By moving the check higher up we avoid this error entirely.

Reviewed By: mgarrard

Differential Revision: D79917972