Work around CV errors due to insufficient data (#4123)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4123

It is possible to fit a TorchAdapter with only one data point but it is not possible to run cross validation with it. This can lead to errors since we rely on CV statistics in `BestModelSelector`, which is used in BOTL. This diff gracefully handles such errors rather than erroring out in the experiment.

Reviewed By: esantorella

Differential Revision: D79915007

fbshipit-source-id: 0dd8a1f5359cee5a12035fd5321c4a28dba2107a