Adapt max_batch_size to model batch shape in optimize_acqf_discrete
Summary: With fully bayesian models, this was leading to quite a large memory usage with the default batch size of 2048. With this change, the max batch size will adapt to the model batch shape, which should bring down the memory usage of batched models to a level similar to non-batched models.

Differential Revision: D80006058