Use floating-point map key types in Benchmark MapMetrics (#4126)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4126

### Context

In the benchmarking set-up, the `normalize_progression=True` option in early-stopping strategies is currently not working properly as zero trials are early-stopped. Specifically, it appears that `is_eligible_any` returns `False` for all trials.

All early-stopping strategies (all subclasses of `BaseEarlyStoppingStrategy`) are impacted. Specifically the instantiation of `MapData` in the return call of `_check_validity_and_get_data` casts the progression column to `int64`, which rounds all the float-valued progressions `[0.0, 1.0)` to `0`, leading `is_eligible_any` to always return `False`

### Changes

This diff removes the specification of `map_key_info` as integer-typed in `BenchmarkMapMetric` and `BenchmarkMapUnavailableWhileRunningMetric`.

Reviewed By: esantorella

Differential Revision: D79938692

fbshipit-source-id: 6e283d8050606c292d2572994aa8925af8c6a477