Adapt max_batch_size to model batch shape in optimize_acqf_discrete (#4127)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4127

With fully bayesian models, this was leading to quite a large memory usage with the default batch size of 2048. With this change, the max batch size will adapt to the model batch shape, which should bring down the memory usage of batched models to a level similar to non-batched models.

Reviewed By: dme65

Differential Revision: D80006058

fbshipit-source-id: 39e17b37678890822afd36f36aad3e72a6bd6c42