clone_reset support for DeterministicGenerationNode (#4134)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4134

clone_reset currently does not work for DeterministicGenerationNode, as the iterator is not reset when the method is called on the generation strategy.

This causes issues when unittesting benchmarking functionality, such as benchmark_replication. Currently, run orchestrator benchmarking tests on a clone of an unused GS, which keeps the issue in this diff from getting exposed. However, This means that we don't have access to the GS's state when computing `BenchmarkResult`s - an issue that is addressed in the subsequent diff.

Differential Revision: D80093006