Updating conditionals for mutually exclusive cases in _overwite_metric (#4145)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4145

### Code Changes

The code changes in `client.py` involve updating the `if` statements to `elif` statements in the `_overwrite_metric` method to ensure that the conditionals are mutually exclusive. This change prevents incorrect metric updates and improves the overall logic of the method.

In `test_client.py`, a new test method `test_overwrite_metric` is added to verify the correct functioning of the `_overwrite_metric` method. The test creates a client, configures an experiment, and tests the replacement of single-objective metrics and outcome constraint metrics.

### Impact

The changes made in this diff improve the reliability and accuracy of the `_overwrite_metric` method, ensuring that the correct metrics are updated in the client.

Reviewed By: mpolson64

Differential Revision: D80005551

fbshipit-source-id: f7aa669de784d45fe6f039f4f7ddae7f36d8780b