MBM: Do not error out if infer_objective_thresholds fails (#4149)
Summary:

`Acquisition._update_objective_thresholds` or `infer_objective_thresholds` call within can lead to errors during `Acquisition` initialization, even if the corresponding acquisition function doesn't need objective thresholds. We have some conditional logic to protect against this but it is not comprehensive enough.

The particular example that triggered this diff is the use of P_Feasible acquisition function when there are no feasible points. In this case, we can't infer objective thresholds since we need feasible points.

This diff adds a `try/except` block to catch errors raised in `infer_objective_thresholds` and logs a warning if any error is caught. This should make it possible to easily debug any related errors (thanks to the log) while letting the acquisition functions that don't need the thresholds run without issues.

Reviewed By: sunnyshen321

Differential Revision: D80259199