Update Winsorize transform to be usable without a config (#4160)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4160

This diff updates `Winsorize` transform to not error out when it encounters a relative outcome constraint when `derelativize_with_raw_sq` option is not set. The transform will now transform the metrics that are not relative and leave the relative ones unchanged (unless `derelativize_with_raw_status_quo` is set to True, in which case they'll also be winsorized).

This will make it possible to use winsorization by default.

Reviewed By: Balandat

Differential Revision: D80361484

fbshipit-source-id: 915a270cade702fea63efce3c3954cd5f0b85ce8