Transition to pyproject.toml for builds (#4100)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4100

The current CI setup that relies on `setup.py` for most build options and very lightly on `pyproject.toml` is broken: https://github.com/facebook/Ax/actions/runs/16761469381/job/47457511203.

I first tried to lightly fix this by marking various fields as dynamic, to allow them to be set by `setup.py`. This was quickly getting out of hand, so I tried a complete migration with the help of devmate.

The setup in this diff mostly replicates the existing `setup.py` but misses one aspect: setting BoTorch pin (`==` vs `>=`) based on `ALLOW_LATEST_BOTORCH` option. I've instead hard-coded it as `>= 0.15.0` which should be ok for the most part. We've added these `ALLOW_LATEST` ones in the past because we had run into some inconsistencies when there were some backwards incompatible changes or a broken release in the underlying dependencies.

Reviewed By: esantorella

Differential Revision: D79678941