Add `STALE` trial status - Part 1 - Update TrialStatus Enum
Summary:
This diff is part 1 of a series of changes to add a new trial status called "STALE" to the Ax library. The new status will be used to indicate trials that have been created but have not yet been run, and may be retried in the future.
The changes involve -
1. Adding 'STALE' as new status to TrialStatus Enum
2. Updating relevant properties and dependencies
3. Added a test file for trial statuses to improve test coverage

Reviewed By: mpolson64

Differential Revision: D80421095