Add `STALE` trial status - Part 1 - Update TrialStatus Enum (#4189)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/4189

This diff is part 1 of a series of changes to add a new trial status called "STALE" to the Ax library. The new status will be used to indicate trials that have been created but have not yet been run, and may be retried in the future.
The changes involve -
1. Adding 'STALE' as new status to TrialStatus Enum
2. Updating relevant properties and dependencies

Reviewed By: mpolson64

Differential Revision: D80421095

fbshipit-source-id: e713af60ac7cab282ab401e9ebb250a0dfcfd1ad